requirements.txt
streamlit
sqlalchemy
pymysql
pandas
python-dotenv
openpyxl
xlrd
cryptography

